// src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      GameSericeKey: {
        Row: {
          created_at: string
          game_id: number
          hash_key: string
        }
        Insert: {
          created_at?: string
          game_id?: number
          hash_key: string
        }
        Update: {
          created_at?: string
          game_id?: number
          hash_key?: string
        }
        Relationships: [
          {
            foreignKeyName: "GameSericeAPI_game_id_fkey"
            columns: ["game_id"]
            isOneToOne: false
            referencedRelation: "GameServiceAdmin"
            referencedColumns: ["game_id"]
          },
        ]
      }
      GameService: {
        Row: {
          created_at: string
          game_id: number
          PlayerInfo: Json | null
          UUID: string
        }
        Insert: {
          created_at?: string
          game_id?: number
          PlayerInfo?: Json | null
          UUID: string
        }
        Update: {
          created_at?: string
          game_id?: number
          PlayerInfo?: Json | null
          UUID?: string
        }
        Relationships: [
          {
            foreignKeyName: "GameService_game_id_fkey"
            columns: ["game_id"]
            isOneToOne: false
            referencedRelation: "GameServiceAdmin"
            referencedColumns: ["game_id"]
          },
          {
            foreignKeyName: "GameService_UUID_fkey"
            columns: ["UUID"]
            isOneToOne: false
            referencedRelation: "UserMaster"
            referencedColumns: ["UUID"]
          },
        ]
      }
      GameServiceAdmin: {
        Row: {
          admin_userid: string
          created_at: string
          game_id: number
          PlayMoney: number
        }
        Insert: {
          admin_userid: string
          created_at?: string
          game_id?: number
          PlayMoney?: number
        }
        Update: {
          admin_userid?: string
          created_at?: string
          game_id?: number
          PlayMoney?: number
        }
        Relationships: []
      }
      QRLogin: {
        Row: {
          id: number
          LoginCode: string
          resultUUID: string | null
        }
        Insert: {
          id?: number
          LoginCode: string
          resultUUID?: string | null
        }
        Update: {
          id?: number
          LoginCode?: string
          resultUUID?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "QRLogin_resultUUID_fkey"
            columns: ["resultUUID"]
            isOneToOne: false
            referencedRelation: "UserMaster"
            referencedColumns: ["UUID"]
          },
        ]
      }
      UserMaster: {
        Row: {
          created_at: string
          Money: number
          UserName: string
          UUID: string
        }
        Insert: {
          created_at?: string
          Money?: number
          UserName: string
          UUID?: string
        }
        Update: {
          created_at?: string
          Money?: number
          UserName?: string
          UUID?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never